comment/move card list to another file
