+++ components/index.ts
@@ -29,6 +29,7 @@ export { default as DeployDid } from "./DeployDid";
 export { default as DIDxWallet } from "./SSI/DIDxWallet";
 export { default as Withdrawals } from "./SSI/DIDxWallet/Withdrawals";
 export { default as Upgrade } from "./SSI/DIDxWallet/Upgrade";
+export { default as CardList } from "./SSI/DIDxWallet/CardList";
 export { default as FAQ } from "./FAQ";
 export { default as SSIProtocol } from "./SSIProtocol";
 export { default as Donate } from "./Donate";
